- [Master] update error handler - [Client] update error handler - [APIGateway] add api update profile - [APIGateway] update all price to double in domain - [APIGateway] update typo at subscription in domain - [APIGateway] add upload photo profile api

+++ Domain/Entities/Requests/Clients/ServicesRequest.cs
@@ -6,6 +6,6 @@
         public int? Duration { get; set; }
         public string? DurationType { get; set; }
         public bool? IsActive { get; set; }
-        public int? Price { get; set; }
+        public double? Price { get; set; }
     }
 }
